Change index format to remove ulid dependency
Seems that using a ulid to make the index keys unique was a dumb strategy,
and using the data record primary key is the standard approach. As such, the
index key is now composed of the key emitted by the application, plus a
separator byte sequence (default is a single null byte), plus the data
record primary key.

API change is an additional separator() function on the Database type, which
permits setting the separator byte sequence to something other than the
default.

cargo test passes